refactor: LottoNumbers제거
